Move log_error to error.py
